+++ src/lib.rs
@@ -31,7 +31,8 @@ Options:
   -V, --version     display version information and exit
 
 Option Parameters:
-  <exp>     regular expression, color the entire match.
+  <exp>     regular expression. Colors the entire match, or the first
+            capture group if one is provided.
 
 Environments:
   AKI_MCOLOR_COLOR_SEQ_RED_ST       red start sequence specified by ansi